10kft - code refactor
